Move dond to submodule
