+++ qcodes/dataset/dond/do_nd.py
@@ -14,11 +14,7 @@ from qcodes import config
 from qcodes.dataset.data_set_protocol import res_type
 from qcodes.dataset.descriptions.detect_shapes import detect_shape_of_measurement
 from qcodes.dataset.descriptions.versioning.rundescribertypes import Shapes
-from qcodes.dataset.experiment_container import Experiment
-from qcodes.dataset.measurements import Measurement
-from qcodes.parameters import ParameterBase
-
-from .do_nd_utils import (
+from qcodes.dataset.dond.do_nd_utils import (
     ActionsT,
     AxesTupleListWithDataSet,
     BreakConditionInterrupt,
@@ -31,11 +27,14 @@ from .do_nd_utils import (
     _register_parameters,
     _set_write_period,
 )
-from .threading import (
+from qcodes.dataset.experiment_container import Experiment
+from qcodes.dataset.measurements import Measurement
+from qcodes.dataset.threading import (
     SequentialParamsCaller,
     ThreadPoolParamsCaller,
     process_params_meas,
 )
+from qcodes.parameters import ParameterBase
 
 LOG = logging.getLogger(__name__)
